+++ public/assets/src/__init__.js
@@ -1,0 +1,2 @@
+import "@/scss/index";
+import "@/js/main";
